+++ wazo_dird/plugins/csv_backend/schemas.py
@@ -8,15 +8,16 @@ from wazo_dird.schemas import BaseSourceSchema
 
 
 class SourceSchema(BaseSourceSchema):
-    db_uri = fields.String(validate=Length(min=1, max=256), required=True)
+    unique_column = fields.String(Length(min=1, max=128), allownone=True, missing=None)
+    file = fields.String(Length(min=1), required=True)
 
 
 class ListSchema(_ListSchema):
 
-    searchable_columns = ['uuid', 'name', 'db_uri']
+    searchable_columns = ['uuid', 'name', 'file']
 
     order = fields.WazoOrder(
-        sort_columns=['name', 'db_uri'],
+        sort_columns=['name', 'file'],
         default_sort_column='name',
     )
     recurse = fields.Boolean(missing=False)